requirements/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
sqlparse==0.5.1
tzdata==2024.1
ruff==0.6.9